Missing dependency in performGitPush on gitTag #87
